add-player-view takes a channel
it shouldn't have to worry about where the heck players is or how to add
things to it. arguably the channel should be part of shared state and
the players-view should be taking things off of it in a go-loop, not the
serra-view. maybe I'll get around to this someday