more styles, time travel
there is one small bug with the time-travel, which is that the state is
updated inconsistenly when the commander damage inc buttons are
pushed. given that this is the most likely use case for these buttons,
we should fix this, but it's very much a PITA. The update to :damages
and :players has to happen as part of the same om/transact!, which is
awkward.